+++ guicey-jdbi/src/main/java/ru/vyarus/guicey/jdbi/installer/MapperInstaller.java
@@ -1,6 +1,8 @@
 package ru.vyarus.guicey.jdbi.installer;
 
 import com.google.inject.Binder;
+import com.google.inject.Binding;
+import com.google.inject.Stage;
 import com.google.inject.multibindings.Multibinder;
 import org.skife.jdbi.v2.tweak.ResultSetMapper;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
@@ -32,15 +34,26 @@ public class MapperInstaller implements FeatureInstaller<ResultSetMapper>,
     }
 
     @Override
-    @SuppressWarnings("unchecked")
-    public <T> void install(final Binder binder, final Class<? extends T> type, final boolean lazy) {
+    public void bindExtension(final Binder binder, final Class<?> type, final boolean lazy) {
         binder.bind(type).in(Singleton.class);
+    }
+
+    @Override
+    public <T> void checkBinding(final Binder binder, final Class<T> type, final Binding<T> manualBinding) {
+        // nothing to do
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    public void installBinding(final Binder binder, final Class<?> type) {
         // just combine mappers in set and special bean, installed by module will bind it to dbi
         Multibinder.newSetBinder(binder, ResultSetMapper.class).addBinding()
                 .to((Class<? extends ResultSetMapper>) type);
 
-        final String target = GenericsResolver.resolve(type).type(ResultSetMapper.class).genericAsString(0);
-        reporter.line("%-20s (%s)", target, type.getName());
+        if (binder.currentStage() != Stage.TOOL) {
+            final String target = GenericsResolver.resolve(type).type(ResultSetMapper.class).genericAsString(0);
+            reporter.line("%-20s (%s)", target, type.getName());
+        }
     }
 
     @Override